+++ CHANGELOG.md
@@ -8,5 +8,6 @@ All notable changes in **python-transip** are documented below.
 - The `transip.v6.objects.InvoiceItemService` service to allow listing all invoice items on a `transip.v6.objects.Invoice` object.
 - The `transip.mixins.ObjectUpdateMixin` mixin to allow calling `update()` on API object directly.
 - Allow an invoice to be written to a PDF file by calling the `pdf()` method on a `transip.v6.objects.Invoice` object.
+- The `transip.v6.objects.ProductService` service to allow listing all products as a `transip.v6.objects.Product` object.
 
 [Unreleased]: https://github.com/roaldnefs/python-transip/compare/v0.3.0...HEAD